+++ AddOns/Auctionator/Source_TBC/Constants/Main.lua
@@ -3,3 +3,8 @@ Auctionator.Constants.Durations = {
   Medium = 24,
   Long = 48,
 }
+
+--FIXME: Added to correct Blizzard error
+if CASTING_BAR_ALPHA_STEP == nil then
+  CASTING_BAR_ALPHA_STEP = 0.05
+end